Put the core repo symlinks back to the 0.10 layout.
This reverts commit 1b01318e0233cc01b2e257220012300fe5be0d59.

There was some confusion here because we pivoted to using the stable-website
branch of core too soon, while it's still tracking 0.9 tags. We need to keep
the submodule tracking master until stable-website is a 0.10 tag, and we
need these symlinks to reflect the 0.10 layout so that we can test the
website against the latest docs.