Move most things into 'content' subdirectory
Having all the submodules in the main directory causes some performance
problems as middleman tries to inotify everything it can find,
particularly on Docker for Mac and Docker for Windows where the disk I/O
performance is poor.

So here we shift all of the things that middleman should look at into
a subdirectory, adding "ext" as a separate bind mount in docker so we
can remain consistent with the expectations of the middleman-hashicorp
docker image.