Link to the Terraform logos for use in READMEs, etc
We have Terraform logos embedded in our READMEs but they currently link
directly to the source repository. To decouple this, we'll expose them on
the main website and link to them from there.

Middleman prunes images that aren't referenced on a page, so this page is
here primarily to cause Middleman to include these images, though it also
serves as a reference for the canonical URLs we should use when writing
READMEs and so forth, presuming that we'll keep these URLs working in
future to avoid breaking the README images again.

This page isn't linked from anywhere specific, but it will show up in the
sitemap and be indexed by search engines, so it needs to remain somewhat
presentable.