WIP: use jquery to identify which sidebar nav sections are active
Unfortunately, this isn't working with the livereload previews, for
reasons I don't fully understand. It looks like it's only firing that
$(document).ready() event once per _tab session,_ which, I've never seen
that before. From a quick stackoverflow search, it looks like probably
what's happening is that it's using ajax to reload the page content
without actually navigating away from the page?