Redirect unique enterprise-legacy pages
This commit adds redirects for every enterprise-legacy page that DOESN'T
encroach on name territory we staked out for the new TFE. This should ensure
that _most_ external links to the old docs don't just fall off the earth;
they'll either redirect, or end up in the new docs if there's a name conflict.
The exceptions are pages that don't have an _exact_ analogue but which are kind
of in the way, like some of the old VCS integration pages. We can either let
those slide or fix them once someone runs into a problem, but to start with I'm
only taking pages where I can be pretty sure we won't accidentally bump into
these as we add more content.