Update TFE SaaS hostname to app.terraform.io
This commit also deletes a longer note about hostnames; now that the new
hostname is visible and official, we don't need any extra explanation of why to
use it for private modules. Hopefully anyone who used private modules in the
intervening weeks is good to go, and new users won't be tempted to use the old
name.