Remove og:image (unfurl image) for now
The current image causes tall unfurls when sharing links to docs in
Slack. Ideally, we could make this image dynamic depending on the
section of the site, but for now I think removing the one we have will
be an overall improvement.

Note that this may affect things like facebook posts that link to the site as
well, but again I think going from "big purple square logo" to "just
title and description and maybe favicon" is AOK.