Move SAML to private TFE section of the sidebar nav
SAML single sign on is only applicable to private TFE, so it should be in with
the private TFE administration and configuration docs. This commit only adjusts
the sidebar nav hierarchy; it doesn't move the saml/ directory, to avoid
breaking links. We can do that later and add redirects if we want, but it seems
unnecessary for now.