sentinel: Update documentation to account for new data namespace
This updates the documentation to reflect the new data source namespace,
which will be available in TFE soon.

The namespace will be available in all 3 imports and is implemented in
the same way resources are. Specifically, this is actually a splitting
of the previously shared "resources" namespace, which will now only
contain resources going forward.