Clarify Sentinel doc link
As a newcomer this link description was weird to me because the Sentinel
documentation doesn't really give a whole lot about enforcing policies
in Terraform, linking right back to this documentation as a more
authoritative reference. Clarify here that the link is about Sentinel
itself and the language, less about Terraform stuff.