Sentinel: Start overhauling the docs
This marks the beginning of the overhaul.

I started with the tfconfig import as we have been working on some other
documentation and we've mapped the namespace for that pretty well. This
first effort is some prelim work necessary to determine what actually
looks good.

At this point in time we might have to go with a more flat layout versus
the per-page detailed layout I wanted, as it looks like there is some
depth limit on indentations on the left sidebar that we've hit. So we
might unroll this into a single page, with adequate detail and
navigation capabilities.