sentinel/import/tfconfig: More doc work, more overhaul
This is a continuation of the documentation for the tfconfig import in
Sentinel/TFE.

I've decided to go back to a singular layout for the import - this
basically means that all of the import documentation will be
self-contained within one page, including all elements of the namespace.

What I've done instead is mapped out the namespace in a tree view at the
top of the document, with details below the map.

This will need to be populated with examples later, which I've decided
to do after I've gone through and documented everything.

There should not be too much work here left to *explain* everything
that's in the namespace, although it might take a bit of time to
construct some basic contextually correct examples after though.