Add TFE<->Sentinel policy examples
This adds direct links to the terraform-guides repository on GitHub.
The links are intentionally set to the file path and not by commit SHA
(as you should normally do when referencing things like this) to not
create another step to updating the website docs when the examples in
the repo are updated.

In the future we may want to do more of an integrated thing to show the
examples inline on the site, but this is a good way to get give the
extremely helpful terraform-guides docs some much needed visibility,
quickly.