docs/sentinel: tfconfig, Move some examples to use null_resource
While I'm keeping some of the examples in the tfconfig import to use
the AWS provider (namely the provider blocks), I want these examples to
be as easy to adopt as possible, and that means less reliance on cloud
resources from one specific vendor.

Since we can provider real executable examples through null_resource,
I've changed some of the ones that are resource-relevant to use that
resource instead of aws_instance.

The provider examples remain the same as the null provider is not
configurable (ie: it lacks config keys to set even).

The variable examples were not applicable as they only contain
variables.