docs/sentinel: tfstate, re-word synopsis
I wasn't happy with the old wording. The entire synopsis seemed like it
was warning the reader off of not using the import, which while might be
good due to its current lack of usefulness (as we only allow policy
checks on plan right now), eventually we don't want that to be the case.

In addition, there was no mention about this is the only namespace you
can get output values right now.

This update:

* Shortens the synopsis greatly, and directs people to the OSS docs for
a deeper explanation of the state.
* Points out that this is the only place outputs can be fetched.
* Relegates any warning of not using the import to an info box.