Add “Creating and Managing Terraform Policies”
This is a new final page in our Enterprise “Getting Started” guide. It adds a walkthrough for adding Sentiel policies to an organization. In this case, we’re just adding a simple passthrough policy that should always pass and not block a run. It references the main Sentinel docs, as well as the other Enterprise Sentinel docs. The content closely follows other pages in the guide, namely the workspaces page.

Signed-off-by: Paul Thrasher <pthrasher@hashicorp.com>