docs/sentinel: Update limitation docs for all mocks
We've discovered that we overlooked the fact that mock count indexes are
represented as strings in the JSON, even though the indexes are integers
in practice.

This can't be necessarily worked around, unfortunately; removing the
indexes and moving the resource objects to an array, for example, will
create edge cases in testing if someone is using a singular-context for
or any/all loop. Hence, we are keeping the mocks the way that they are,
and just noting the limitations.

I've also revised the limitations so that they are not info boxes, but
but rather simple bullet points. tfconfig, even though it's not affected
by this, has been re-formatted as well to stay consistent.