Reduce the user impersonation API documentation
The workflow for using this is gnarly enough that 1: the existing documentation
was incomplete, and 2: the documentation should encourage you to not use it. But
in recognition that we can't anticipate every possible use case, we'll leave in
a note to ask for support if you run into a case where it is absolutely
necessary.