Layouts: Large, on-brand link preview image for marketing pages
In the name of good Slack citizenship, we need to keep preview images small when
users and employees paste docs links into chats, and 95% of the content on this
site is docs. But the marketing pages should use whatever the marketing and design
teams deem brand-appropriate! This commit updates the layouts to enable that.

So far, we've identified the following pages as marketing content:

- /index.html
- /logos.html
- /intro/* (requires separate PR to terraform repo)