Remove bitbucket cloud from OAuth creation docs
Bitbucket does not support using personal access tokens as oauth tokens, so our API to create a client was incorrect. This was confusing users. Bitbucket also does not issue oauth tokens via their API (just the normal OAuth flow), so we cannot easily code a workaround that fetches a token.