Add symlink to new Tf CLI docs front page
This page lives in the terraform repo, and acts as the front page for
all the /docs content that comes from that repo.

I recognize that it's somewhat annoying to have a front page named
something other than index.html. But it's somewhat more annoying to
have to move everything into a whole new subfolder, lay down 301s, and change
EVERY LINK across multiple repos accordingly.