Link to provider docs in sidebar quick links
Evan reminded me that the people who use the site _every day_ are probably going
straight to the provider docs, and they deserve a shortcut. We might do more
later to surface the provider docs (and/or separate them from the rest of the
CLI docs), but this will at least make sure they have a one-touch path from any
docs page.

This commit also moves the glossary a little lower in the list, and removes the
downloads page from the list (it's already in the top nav, and it's not properly
"docs," so it eventually felt weird to have there.)