Update hashicorp-middleman container to v0.3.39
This updates some dependencies with known issues. It does not appear to change
any website behavior.

https://github.com/hashicorp/middleman-hashicorp/compare/v0.3.35..v0.3.39

NOTE TO FUTURE MAINTAINERS: When upgrading the container version, you should
also update the Gemfile.lock to ensure it matches the gems currently installed
in the container. This makes builds and local previews faster by skipping some
extra network and install activity. To do this:

- Delete content/Gemfile.lock.
- Run `make website` once, and wait until Middleman comes online.
- Kill the container.
- Commit the updated content/Gemfile.lock.