Collapse the other nav items into an 'info' menu at problematic widths
Now that the nav bar is wider, there's a point somewhere in between "large" and
"mobile" sizes where it would overlap the logo image, which causes it to get
kicked down a line. That looks janky. One fix would be to switch to the
hamburger menu earlier, but I think it's more functional to have an in-between
state where you can still use the nav bar as a bar.