Add brief rotation effect when subnavs change state
I couldn't find the glyph I wanted for down-chevrons, but that's OK, because
faking it with the `rotate()` function has the fringe benefit of letting us
animate the transition.

When you're directly interacting with the toggles, that's just decorative. But
it also auto-animates on initial page load because of the order in which the
classes get switched, and THAT's actually functional because it's a signal that
the chevrons are interactable. (They don't auto-animate on back/forward
navigation, since the classes are already there on the turbolinks cached
versions of those elements.)