Add "jump to content" link on mobile
This link is only visible if the viewport is narrow enough that the nav sidebar
collapses. At that point, you need to scroll past the entire nav content before
you reach the page content, which makes the site very tiring on mobile.

Having to tap an extra link before reading still isn't the best of experiences,
but it's nicer than scrolling all that way.