Increase text contrast
Users have sometimes pointed out that the text contrast on terraform.io
is low. More recent updates to our web properties use text colors in
the #3x3x3x sort of range (for example, vaultproject.io uses #373942),
which feels like a better fit for us as well.

Changing $gray-darker affects a few more things than body text:

- The callout boxes on the home page.
- The header links on the home page.
- The sidebar links on docs pages.

IMO this is all to the good; it reduces the intensity of the hover effects
(which go to `#000000`), but I don't mind that.