sentinel: update mock generation documentation
This commit does the following:

* Adds a whole new section on mock generation, which details how to
get the information from the UI or API, and how to lay out the data
in Sentinel testing scenario.

* Updates the existing Sentinel documentation. We either refer to the
new documentation (in most of the general guide cases), or just
remove references to old detail completely (mostly in the case of the
"mocking tf[config|plan|state] data" pages, which have now been
deleted).

* Makes a couple of other quality-of-life changes such as re-ordering
the import list alphabetically in the sidebar, and updating the index
page with details on every main section, which I don't think had been
done in a bit.

Co-Authored-By: Chris Arcand <chris@chrisarcand.com>
Co-Authored-By: Alexis Grant <alexisg@gmail.com>
Co-Authored-By: Nick Fagerlund <nick@hashicorp.com>