change --no-tty to --tty=0
Using --no-tty in `replicated admin --no-tty retrieve-iact` causes the command to hang.  Doing this in a shell is not too problematic because the user can just use <ctrl>-c to get their prompt back.  But this causes problems in scripts.  Using `--tty=0` is better and is in fact what Replicated recommends for scripts here: https://help.replicated.com/docs/swarm/packaging-an-application/admin-commands/#cli-script-flags