extend: Documentation for provider developers about v0.12 compatibility
This is a first pass at documentation for upgrading providers for
compatibility with Terraform Core v0.12. I expect we'll do further
iterations on this as we do more releases and learn more, but the goal
here is to get some initial information out there ASAP to give some runway
to get as many providers as possible updated before v0.12 final.