New functions to interate over modules in tfplan
The original example function that showed how to iterate over modules in the tfplan import was inadequate in that it did not include the name of the resource or the module in which it was found in the list it returned.  I have defined a generic, parameterized function, `find_resources_from_plan` that accepts the resource type as an argument. 

This function can be used unchanged in all policies that use the tfplan import. Instead of returning a list of maps, it returns a map of maps in which the keys are the module path and the values are maps containing all resources of the specfied type in that module. This allows policies that call it to output the module name and full resource path in calls to the Sentinel `print()` function, making the outputs from TFE Sentinel policies much clearer to users when their plans violated Sentinel policies.

Additionally, I've defined an auxillary function, `get_address`, that returns the address of a resource using the form `module.A.module.B.<type>.<name>` that is used in plan and apply logs. I also show how to get the full address of a resource instance by appending something like `"[" + string(index) + "]"`.

You can see examples of these function being used in this [policy](https://github.com/rberlind/new-policies/blob/master/aws/restrict-ec2-instance-type.sentinel) and others in the same repository. The versions of the functions in that repository have additional comments, but I left them out of the text in this page since the surrounding text explains the functions, making the comments unnecessary here.  In the near future, I plan to merge the policies in that repository into the official [hashicorp/teraform-guides/governance](https://github.com/hashicorp/terraform-guides/tree/master/governance) repository of official Sentinel examples.

I am also updating the doc pages for the tfconfig and tfstate imports to reference similar functions as part of the same PR.