New functions to interate over modules in tfconfig
The original example function that showed how to iterate over modules in the tfconfig import was inadequate in that it did not include the name of the resource or the module in which it was found in the list it returned.  I have defined a generic, parameterized function, `find_resources_from_config` that accepts the resource type as an argument. 

This function can be used unchanged in all policies that use the tfconfig import. Instead of returning a list of maps, it returns a map of maps in which the keys are the module path and the values are maps containing all resources of the specfied type in that module. This allows policies that call it to output the module name and full resource path in calls to the Sentinel `print()` function, making the outputs from TFE Sentinel policies much clearer to users when their plans violate Sentinel policies.

Additionally, I've defined an auxillary function, `get_address`, that returns the address of a resource using the form `module.A.module.B.<type>.<name>` that is used in plan and apply logs.

I have already updated the doc page for the tfplan import to reference similar functions and will soon do the same for the tfstate import and submit all 3 as part of the same PR.