New functions to interate over modules in tfstate
The original example function that showed how to iterate over modules in the tfstate import was inadequate in that it did not include the name of the resource or the module in which it was found in the list it returned.  I have defined a generic, parameterized function, `find_resources_from_state` that accepts the resource type as an argument. 

This function can be used unchanged in all policies that use the tfstate import. Instead of returning a list of maps, it returns a map of maps in which the keys are the module path and the values are maps containing all resources of the specfied type in that module. This allows policies that call it to output the module name and full resource path in calls to the Sentinel `print()` function, making the outputs from TFE Sentinel policies much clearer to users when their state violates Sentinel policies.

Additionally, I've defined an auxillary function, `get_address`, that returns the address of a resource using the form `module.A.module.B.<type>.<name>` that is used in plan and apply logs. I also show how to get the full address of a resource instance by appending something like `"[" + string(index) + "]"`.

I have also updated the doc pages for the tfplan and tfconfig imports to reference similar functions as part of the same PR.