extend: Split Resources page into separate pages
The single page content was very long and we have planned additions to various sections of it. To split this into its own section, we create a new subfolder, convert the existing page into an index page, and setup a redirect for the previous URL.

The content on the section subpages is a lift and shift of the existing content while the index page follows our typical model of introducing each subpage. The sections of introductions on the index page were slightly reordered to reflect more common use cases.