Sidebars: New class for sub-navs: `nav-auto-expand`
If you add the `nav-auto-expand` class to a subnav `<ul>`, that section will
default to active _without_ activating its parents and grandparents. This means
that the section in question will be visible as soon as you expand its immediate
parent.

This is for cases where we benefit from dividing sub-items into groups, but we
_don't_ benefit from hiding the children of those groups behind an additional
click.

Canonical example: AWS provider maintainers wanted to group their resources and
data sources into related categories, like this:

- Autoscaling
    - Data Sources
        - `aws_autoscaling_group`
        - ...
    - Resources
        - `aws_autoscaling_attachment`
        - ...

They wanted both of those sub-items to default to expanded once "Autoscaling"
was expanded, because:

- The user might want any of the items in either category.
- The user probably wants to see the full listings of both categories before
  making further decisions.
- The sizes of the categories were quite manageable, by AWS standards.
- Requiring an additional click would be obnoxious.