on-premise to self-hosted
It is incorrect to describe TFE as "on-premise".  While it is on-premise when installed on VMware, it is often installed in virtual private networks (VPCs) in AWS, Azure, and Google Compute Cloud.  So, it is better to use the term "self-hosted" which is also what we use on https://www.hashicorp.com/products/terraform/pricing