CSS / JS: Use CSS grid to line up main headers and controls
The way I was doing this before was super brittle. This should be a lot more
resilient.

It does require using JS to move the headers and controls into container divs.
That's because the main page title and the sidebar title are just embedded in
the markup, rather than in structured data of some kind.