CSS: Fix cut-off resource page titles in sidebar
We were using `display: table-cell` to keep the bullets aligned with the text and
provide a nice hanging-indent effect for multi-line titles, but unfortunately
it's impossible to force mid-word wrapping when using `display: table-cell`, so
long _single-word_ titles (usually resources or data sources in the provider
docs) would just overflow into the abyss.

But you can get the same alignment effect with `display: grid`, and it offers
much better control over max sizes. This commit styles sidebar list items as
little four-cell grids, with the sub-nav (if any) taking the bottom two cells.
It also leaves the existing table-cell solution in place for old browsers that
fail a `@supports (display: grid)` query. (So nothing gets worse, but we only
fix the problem on platforms where it's fixable.)