Fix Travis failures by specifying exact Bundler version
Some update to the Travis worker images (at v6.2.6) changed the default Bundler
behavior for our version of Ruby; it now defaults to 2.1 instead of 1.x.
Naturally that blows everything up, because our ancient Middleman relies on
Bundler 1.x:

```
Resolving dependencies...
Bundler could not find compatible versions for gem "bundler":
  In Gemfile:
    middleman-hashicorp (= 0.3.41) was resolved to 0.3.41, which depends on
      middleman (~> 3.4) was resolved to 3.4.0, which depends on
        middleman-core (= 3.4.0) was resolved to 3.4.0, which depends on
          bundler (~> 1.1)
  Current Bundler version:
    bundler (2.1.0)
```

Just installing the proper version of Bundler won't work, and uninstalling 2.1
requires a truly heinous hack because it's a "default gem." But the `bundle`
binary actually lets you specify which version of the Bundler gem it should
load, and changing the install command to do that seems to work fine, so we get
to kick the can down the road for a few more months.