cloud: add Sentinel tfconfig/v2 import documentation
This adds the import documentation for tfconfig/v2, the next-generation
Terraform Sentinel import for config introspection.

Note that the documentation is currently light on examples. Initially,
the documentation will not include too much in the way of examples, as
I'd like to get some more functional (while still being terse)
real-world examples in for each collection rather than the kind of
"spec" style examples that you see in the current Sentinel imports. This
will eventually segue into live examples that we will back with a future
Sentinel playground.

Note that some links will be broken until the docs for tfplan/v2 and
tfstate/v2 are completed. This will not be merged before those are
complete.