cloud: add Sentinel tfplan/v2 import documentation
This adds the import documentation for tfplan/v2, the next-generation
Terraform Sentinel import for plan introspection.

The format follows a similar layout to tfconfig/v2 (see #1090). We
document the collections, and go into detail on how to work with change
representations. The raw "collection" is also called out as a method to
access low-level plan details outside of the structure of the
collections.