cloud: remove Sentinel tfstate message mentioning future work
This has been part of the Sentinel documentation for a long time and we
have not really determined a way to apply policy checks to anything else
other than the plan at this point in time. Not to say that this will
never happen, but we probably should not be implying that it will happen
indefinitely.

This just re-purposes the v2 info box for the v1 imports - basically
suggesting that someone can use "applied" to the same effect.