Guts: Fix the "Edit this page" link
This feature broke when we switched to CircleCI for site builds, but it took us
enough time to notice the problem that we didn't immediately associate it with
the build changes.

Anyway, this method depended on the exact layout of the container's filesystem
root when invoking Docker in a particular way, and Circle doesn't set up the
volumes like that. So let's make it more resilient. It still depends on the
relative layout of these directories, but that's what the symlinks themselves
rely on, so it seems like a good boundary for assumptions.

Adding this additional path math on every page might cause a build time
performance hit, so we should check the average build times a little while after
this ships to make sure it's not too obnoxious. (If so, we can do the path math
more efficiently; it'll just be less legible and more error-prone.)