cloud: Remove approximate address language from tfconfig/v2
This removes the "approximate address" language from the tfconfig/v2
import documentation.

We've decided that this language is not really serving any purpose.
Going forward, these are just "addresses", just without indexes if
expansion is a thing.

The docs have been updated to document this difference between the
addresses found in tfconfig/v2 and tfplan|tfstate/v2.

Additionally, we're renaming the approximate_address helper to
strip_index. This is also reflected in the docs.