cloud: add section on configuring Sentinel modules
This adds a small section for configuring modules to the policy
management documentation.

Nothing too fancy or meaty about it - just how to add it to the
sentinel.hcl file and a callout to the fact that you can't include
modules outside of the policy set working directory hierarchy right now.