adding permissions information
per TOB-TFE-036 in the recent TFE security assessment, it was noted that we are not explicit that permissions between VCS and TFC/E are not synchronized. This PR adds information to indicate that customers should be aware of the security implications of using any given VCS user as a conduit between these two systems.