JS: Fix bug with multiple sidebar headers
If a sidebar had multiple headers (besides the special #otherdocs one), they
could sometimes end up superimposed on each other, depending on the state
of the page when 'turbolinks_load' fired. This was because the initial header
was no longer matched by '.children' after moving to the grid container.
Now it stays matched, so the script won't try to move anything else into its place.