sentinel: update module and policy configuration details
Monorepo support and remote sources for modules have been introduced
for Sentinel in TFC. It's now possible to use module and policy sources
in a uniform fashion - both now allow local and remote sources, and the
new monorepo support means you can now include modules from a shared
directory if you so choose.

In that light, some of the documentation needed updating to remove
mention of any restrictions relating to monorepos or the ability to
include policy sources that are local files. This makes those changes.

It also ensures policy block examples include the "source" field now
uniformly, as we are deprecating the optional nature of this field
within policies - eventually, it will be a required field, just like
modules.