Cover exposure of sensitive variables in Sentinel mocks
I updated the text related to sensitive variables to cover Sentinel mocks.
I also changed the old language "Marking a variable as sensitive prevents anybody (including you) from viewing its value in Terraform Cloud's UI or API." to "Marking a variable as sensitive prevents anybody (including you) from viewing its value in the variables section of the workspace in Terraform Cloud's UI or with its Variables API endpoint." which is more accurate.  The first was NOT accurate because people with the right permission CAN view sensitive Terraform variables in state files in the UI if they are passed to attributes of resources. Additionally, someone with full TFC API access could extract the state file and again see values of sensitive Terraform variables in that case.