redirect removed examples pages
The examples section of terraform's website had pages which were
pointing directly to examples in the terraform repository. Those
examples were very out of date and I removed them without realizing that
they were references on the website.

There is better content on learn.hashicorp.com, so we decided to remove
the Examples section entirely and replace with a general redirect to
learn.