JS: Re-enable turbolinks navigation in sidebar navigation links
Stopping propagation must have been a leftover from when I was developing this
and doing something slightly different in the handlers -- it's not needed, and
it causes sidebar navigation links to trigger a complete page load instead of a
fast turbolinks transition.

Prior to this commit, top-level sidebar nav links (and normal links in body
text) were fast, but nested sidebar nav links were slow.