JS: Fix a variety of nasty turbolinks interactions
The `turbolinks:load` event is problematic if you're modifying the page layout
and also navigating to anchor links, because turbolinks manually modifies your
scroll position right before firing its load event. Since a requested scroll is
async and the functions for it aren't thennable, it turns into a race condition
with the subsequent page munging, resulting in unpredictable and frustrating
behaviors. Among other things, this was making the FIRST in-page anchor click to
a header misfire, and was sometimes making it impossible to link to an
auto-linked list item on initial page load.

So instead of using the load event, we'll munge the HTML on:

- Initial page load, which only happens once.
- The `turbolinks:before-render` event, which gives us access to a fully prepped
  body element and lets us munge it _before_ turbolinks tries to figure out
  where it's scrolling to.

In order to deal with the detached body element in before-render, we need to
refactor our event handler to pass the appropriate body element as a context
argument whenever it uses jQuery to select elements in the page.