Deploys: Split into upload and warm-cache steps, keep wrapper script
Our current website deployment CI doesn't distinguish between broken links
(trivial, to be fixed at our convenience) and hosed builds (apocalyptic,
preempts everything). So, let's fix that by displaying their results as separate
steps in the deploy workflow. First step is to split the deploy script to enable
that.

To avoid breaking anything that currently calls the deploy.sh script, this
preserves that script as a wrapper that calls both steps in sequence.