CircleCI: Reconfigure Slack notifications with orb
Circle's built-in Slack notifications (which are apparently deprecated anyhow)
only notify once at the end of the workflow, so we don't get the benefit from
splitting the deploy linkcheck jobs -- it's always just one cryptic red block in
our channel.

But the Slack orb is more configurable, and lets us give more informative
messages on a per-job basis! In fact, we can even do _multiple_  messages per
job, which will be nice for letting people know that their deploy is WIP.