Remove MOST provider submodules
As of December 18, 2020, we've redirected nearly all of the provider
documentation. For providers that never got adopted by a new publisher, we
archived the GitHub repository and redirected each docs page to the
corresponding Markdown source file on github.com. (For an example of these
redirects, see https://www.terraform.io/docs/providers/telefonicaopencloud/r/s3_bucket.html)

This commit removes the submodules, layout symlinks, docs folder symlinks, and
incoming-links.txt entries for every provider that we've redirected on prod.
Namely, the following:

- arukas
- azure
- bitbucket
- cherryservers
- clc
- cloudstack
- cohesity
- dyn
- hedvig
- ignition
- incapsula
- influxdb
- jdcloud
- librato
- logentries
- logicmonitor
- metalcloud
- netlify
- oneandone
- powerdns
- pureport
- rancher
- rightscale
- runscope
- softlayer
- statuscake
- telefonicaopencloud
- ultradns

It leaves ten provider submodules in place, since those are gonna get published
on the registry but still need a little bit of time:

- avi
- chef
- cobbler
- genymotion
- github
- infoblox
- ksyun
- mysql
- rubrik
- rundeck