Fix text munging in incoming links check
The old incoming links check includes logic for ignoring files that already got
redirected. Great!

However, the `grep -v | sort | uniq` command was leaving a single empty line in
the list of redirected paths. Grep considers an empty line to be a pattern that
matches every string. So when the *second* `grep -v` command tried to determine
"which incoming links *aren't* redirected" by letting through any paths that
didn't match any lines in the redirects list... it was getting none of them,
because every path matched against that empty line.

I think there was a point where this wasn't a problem, because the redirects
file didn't include any empty lines? Or something.

So, remove empty lines (with or without spaces) from the list of redirects at
the same time we remove comments.