Convert old check-links.sh script to check-incoming-links.sh
The new link checking jobs spider the site and do spot checks on PRs, but they
don't check the list of pages that we know other sites link to, which is sort of
a fallback to make sure we're not breaking the rest of the web even if we
updated our own stuff.

So, this commit breaks that task out into its own independent script, and
deletes the rest of the old wget-based spidering link checker.