CI: Add job+workflow to check incoming-links list weekly on Mondays
This job is a little costly to run on deploys (since, if we want to keep the
deploy success/failure notification separate from the link check notification,
we would nede to spin up the middleman container one more time per deploy), but
we still want to run it periodically. So, a weekly schedule seems like a good
call.