PR link check: add special casing for Vercel routes
This should help avoid false positives for the path prefixes that get served by
the Next/Vercel app in production. Unfortunately there's not a good way to
discover which paths that currently applies to, but at the time of writing it's
just /cloud and /community, and by the time they expand it much further, we'll
probably be looking to throw this script away anyhow.